requirements.txt
python-binance
pandas
numpy
matplotlib
ta-lib
dotenvdotenv
os